[IMP] hr_employee_firstname: rename hooks file
Apply guidelines on name of hooks file.
