+++ hr_employee_firstname/__init__.py
@@ -1,4 +1,4 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 from . import models
-from .init_hook import post_init_hook
+from .hooks import post_init_hook
