+++ hr_employee_firstname/__manifest__.py
@@ -8,7 +8,7 @@
     "Fekete Mihai (Forest and Biomass Services Romania), "
     "Onestein, "
     "Odoo Community Association (OCA)",
-    "maintainer": "Savoir-faire Linux",
+    "maintainers": ["Savoir-faire Linux", "luisg123v"],
     "website": "https://github.com/OCA/hr",
     "license": "AGPL-3",
     "category": "Human Resources",
